Split native Zenmap output action helpers
