+++ zenmap/macos/native/Zenmap.swift
@@ -219,21 +219,21 @@ struct ContentView: View {
     @State var newProfileDescription = "Custom scan profile."
     @State var selectedProfileID: ScanProfile.ID?
     @State var profileFilterText = ""
-    @State private var output = "Ready. Choose a profile, enter a target, then run a scan."
-    @State private var status = "Idle"
-    @State private var exitStatus: Int32?
+    @State var output = "Ready. Choose a profile, enter a target, then run a scan."
+    @State var status = "Idle"
+    @State var exitStatus: Int32?
     @State var isRunning = false
     @State var selectedTab = "Output"
     @State private var baselineCompareScanID: SavedScan.ID?
     @State private var comparisonCompareScanID: SavedScan.ID?
     @State private var isOutputFindVisible = false
     @State private var isOutputAutoScrollEnabled = true
-    @State private var outputFindText = ""
-    @State private var outputFindSelection = 0
+    @State var outputFindText = ""
+    @State var outputFindSelection = 0
     @FocusState private var isOutputFindFocused: Bool
     
     @State private var runningProcess: Process?
-    @State private var privilegedScanPID: Int32?
+    @State var privilegedScanPID: Int32?
     @State private var privilegedChildPIDPath: String?
     @State var scanStartedAt: Date?
     @State var scanProgressPercent: Double?
@@ -249,8 +249,8 @@ struct ContentView: View {
     @State private var pendingScanOutputBuffer = ""
     @State private var pendingScanProgressBuffer = ""
     @State private var pendingScanOutputFlushWorkItem: DispatchWorkItem?
-    @State private var lastCommand = ""
-    @State private var lastXMLPath = ""
+    @State var lastCommand = ""
+    @State var lastXMLPath = ""
     
     @State var hosts: [ScannedHost] = []
     @State private var selectedHostID: ScannedHost.ID?
@@ -3259,92 +3259,6 @@ struct ContentView: View {
         nseScriptHelperMessage = "Copied profile arguments to clipboard."
     }
 
-    private func copyOutput() {
-        NSPasteboard.general.clearContents()
-        NSPasteboard.general.setString(output, forType: .string)
-    }
-
-    private func copyDiagnosticInfo() {
-        NSPasteboard.general.clearContents()
-        NSPasteboard.general.setString(diagnosticInfoText(), forType: .string)
-        output += "\nCopied diagnostic info to clipboard."
-    }
-
-    private func diagnosticInfoText() -> String {
-        let nmapPath = nmapBinaryPath() ?? "unavailable"
-        let nmapDirectory = nmapPath == "unavailable" ? "unavailable" : nmapDataDirectory(for: nmapPath)
-        let bundlePath = Bundle.main.bundlePath
-        let appVersion = Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String ?? "unknown"
-        let appBuild = Bundle.main.object(forInfoDictionaryKey: "CFBundleVersion") as? String ?? "unknown"
-        let processInfo = ProcessInfo.processInfo
-        let macOSVersion = processInfo.operatingSystemVersionString
-        let hostName = Host.current().localizedName ?? Host.current().name ?? "unknown"
-        let nmapVersion = nmapVersionText(nmapPath: nmapPath, nmapDirectory: nmapDirectory)
-
-        return [
-            "Zenmap Diagnostic Info",
-            "",
-            "App:",
-            "Bundle: \(bundlePath)",
-            "Version: \(appVersion)",
-            "Build: \(appBuild)",
-            "",
-            "System:",
-            "macOS: \(macOSVersion)",
-            "Host: \(hostName)",
-            "Architecture: \(processInfo.machineHardwareName)",
-            "",
-            "Nmap Runtime:",
-            "Nmap binary: \(nmapPath)",
-            "NMAPDIR: \(nmapDirectory)",
-            "Nmap version:",
-            nmapVersion,
-            "",
-            "Last Scan:",
-            "Command: \(lastCommand.isEmpty ? "none" : lastCommand)",
-            "XML: \(lastXMLPath.isEmpty ? "none" : lastXMLPath)",
-            "Exit status: \(exitStatus.map(String.init) ?? "none")",
-            "Status: \(status)",
-            "Hosts parsed: \(hosts.count)",
-            "Ports parsed: \(allPorts.count)",
-            "",
-            "Privilege:",
-            "Currently running: \(isRunning ? "yes" : "no")",
-            "Privileged scan PID: \(privilegedScanPID.map(String.init) ?? "none")"
-        ].joined(separator: "\n")
-    }
-
-    private func nmapVersionText(nmapPath: String, nmapDirectory: String) -> String {
-        guard nmapPath != "unavailable",
-              FileManager.default.isExecutableFile(atPath: nmapPath) else {
-            return "unavailable"
-        }
-
-        let process = Process()
-        process.executableURL = URL(fileURLWithPath: nmapPath)
-        process.arguments = ["--version"]
-
-        var environment = ProcessInfo.processInfo.environment
-        if nmapDirectory != "unavailable" {
-            environment["NMAPDIR"] = nmapDirectory
-        }
-        process.environment = environment
-
-        let pipe = Pipe()
-        process.standardOutput = pipe
-        process.standardError = pipe
-
-        do {
-            try process.run()
-            process.waitUntilExit()
-            let data = pipe.fileHandleForReading.readDataToEndOfFile()
-            return String(data: data, encoding: .utf8)?
-                .trimmingCharacters(in: .whitespacesAndNewlines) ?? "unavailable"
-        } catch {
-            return "unavailable (\(error.localizedDescription))"
-        }
-    }
-    
     private func saveCurrentXML() {
         guard !lastXMLPath.isEmpty else {
             output += "\nNo XML scan result is available to save."
@@ -3413,17 +3327,6 @@ struct ContentView: View {
         }
     }
 
-    private func printOutput() {
-        let printView = NSTextView(frame: NSRect(x: 0, y: 0, width: 720, height: 960))
-        printView.string = output
-        printView.isEditable = false
-        printView.font = NSFont.monospacedSystemFont(ofSize: 10, weight: .regular)
-
-        let printOperation = NSPrintOperation(view: printView)
-        printOperation.jobTitle = lastCommand.isEmpty ? "Nmap Scan Output" : lastCommand
-        printOperation.run()
-    }
-    
     private func openXML() {
         let panel = NSOpenPanel()
         panel.title = "Open Nmap XML"
